Create lampturnin
This code is designed to control a lamp using an Arduino microcontroller. Here's what it does:

Variable Declaration:

It declares an integer variable lampPin, which represents the digital pin number to which the lamp is connected. In this example, it's assumed that the lamp is connected to pin 13.
Setup Function (setup()):

It initializes the serial communication with a baud rate of 9600 baud using Serial.begin(9600).
It configures the lampPin as an output pin using pinMode(lampPin, OUTPUT).
Loop Function (loop()):

It continuously checks if there is any data available to read from the serial port using Serial.available().
If there's data available, it reads the incoming command character using Serial.read() and stores it in the variable command.
If the received command is '1', it turns on the lamp by setting the lampPin high (digitalWrite(lampPin, HIGH)), and sends a message "Lamp turned ON" to the serial monitor.
If the received command is '0', it turns off the lamp by setting the lampPin low (digitalWrite(lampPin, LOW)), and sends a message "Lamp turned OFF" to the serial monitor.
Serial Communication:

The code expects commands to be sent via the serial port. When you send a command (either '1' or '0') to the Arduino via the serial monitor or any other serial communication program, it will turn the lamp on or off accordingly.
Overall, this code sets up an Arduino to listen for commands ('1' to turn the lamp on and '0' to turn it off) sent through the serial port, and it controls the lamp based on those commands.